requirements.txt
certifi
flet
pyyaml
flet-audio
flet-lottie